changed all references to io to new name and structure. Also changed references to sd in tests.
